Stacky Agents/backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
SQLAlchemy==2.0.36
alembic==1.13.3
pydantic==2.9.2
python-dotenv==1.0.1
python-json-logger==2.0.7
requests==2.32.3
truststore==0.10.4
PyYAML==6.0.3
pytest==8.3.3
pytest-flask==1.3.0
